Refactor for Dijkstra
