+++ src/main/java/interpreter/BfsSnapshot.java
@@ -1,0 +1,14 @@
+package interpreter;
+
+import model.Node;
+
+import java.util.LinkedList;
+import java.util.Queue;
+
+public class BfsSnapshot extends SearchSnapshot<Queue<Node>> {
+    public BfsSnapshot() {
+        toTraverse = new LinkedList<>();
+        nextNodeProvider = toTraverse::remove;
+        nodeAdder = toTraverse::add;
+    }
+}
