+++ src/main/java/interpreter/InterpreterException.java
@@ -1,3 +1,5 @@
+package interpreter;
+
 public class InterpreterException extends RuntimeException {
     public InterpreterException(String message) {
         super(message);
