+++ src/main/java/interpreter/SearchSnapshot.java
@@ -1,0 +1,19 @@
+package interpreter;
+
+import lombok.Data;
+import model.Node;
+
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.function.Consumer;
+import java.util.function.Supplier;
+
+@Data
+public class SearchSnapshot<T extends Collection<Node>> implements Snapshot {
+    Set<Node> visited = new HashSet<>();
+    T toTraverse;
+    Supplier<Node> nextNodeProvider;
+    Consumer<Node> nodeAdder;
+    Node current;
+}
