+++ src/main/java/interpreter/Snapshot.java
@@ -1,0 +1,4 @@
+package interpreter;
+
+public interface Snapshot {
+}
