+++ src/main/java/parser/ParsingException.java
@@ -1,3 +1,5 @@
+package parser;
+
 public class ParsingException extends RuntimeException {
     public ParsingException(String message) {
         super(message);
